add get-location.js
